+++ apps/admin/src/main.tsx
@@ -1,5 +1,4 @@
 import { StrictMode } from 'react';
-import * as ReactDOM from 'react-dom';
 import { createRoot } from 'react-dom/client';
 
 import { App } from './app/app';