+++ libs/nestjs/authentication/src/interfaces/index.ts
@@ -1,4 +1,1 @@
-export * from './authentication-module.interface';
-export * from './password-service.interface';
-export * from './user.interface';
-export * from './user.service.interface';
+export * from './authentication-module-options.interface';