+++ libs/nestjs/bcrypt/src/interfaces/encryption.service.ts
@@ -1,0 +1,4 @@
+export interface EncryptionService {
+  hash(value: string): Promise<string> | string;
+  compare(value: string, hash: string): Promise<boolean> | boolean;
+}