+++ libs/nestjs/password/src/constants/reset-password.constant.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_RESET_PASSWORD_LINK = `/password/reset/{{id}}/{{code}}`;