+++ libs/nestjs/winston/src/winston-nestjs-logger.format.ts
@@ -34,7 +34,9 @@ export const nestLikeConsoleFormat = (
     const color =
       nestLikeColorScheme[level] || ((text: string): string => text);
 
-    const stringifiedMeta = safeStringify(meta);
+    const { stack = '', ...rest } = meta;
+
+    const stringifiedMeta = safeStringify(rest);
     const formattedMeta = prettyPrint
       ? inspect(JSON.parse(stringifiedMeta), {
           colors: !!colors,
@@ -44,12 +46,17 @@ export const nestLikeConsoleFormat = (
 
     const name = color(`[${appName}]`);
     const levelLabel = level.charAt(0).toUpperCase() + level.slice(1);
-    const timeLabel = typeof timestamp !== 'undefined' ? `${timestamp} ` : '';
+    const timeLabel =
+      typeof timestamp !== 'undefined'
+        ? `${timestamp ? `${timestamp} ` : ''}`
+        : '';
     const contextLabel =
       typeof context !== 'undefined' ? `${clc.yellow(`[${context}]`)} ` : '';
 
     return `${name} ${levelLabel}\t${timeLabel}${contextLabel}${color(
       message,
-    )} - ${formattedMeta}`;
+    )}${formattedMeta === '{}' ? '' : ` - ${formattedMeta}`}${
+      stack.length > 0 ? `\n${color(stack)}` : ''
+    }`;
   });
 };